test(gemini): split runtime route regressions
